+++ src/main/java/coz/weavon/common/application/model/command/RestCommand.java
@@ -1,7 +1,7 @@
-package coz.weavon.common.application.model;
+package coz.weavon.common.application.model.command;
 
+import coz.weavon.common.application.model.exception.BusinessException;
 import coz.weavon.common.application.util.DateTimeUtils;
-import coz.weavon.common.presentation.model.exception.BusinessException;
 import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.Map;
