+++ src/main/java/coz/weavon/common/application/model/exception/ClientException.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.presentation.model.exception;
+package coz.weavon.common.application.model.exception;
 
 public class ClientException extends RestException {
     public ClientException(String message) {
