+++ src/main/java/coz/weavon/common/presentation/validator/PaginationValidator.java
@@ -1,6 +1,6 @@
 package coz.weavon.common.presentation.validator;
 
-import coz.weavon.common.presentation.model.exception.ClientException;
+import coz.weavon.common.application.model.exception.ClientException;
 import coz.weavon.common.presentation.model.reqeuest.PageRequest;
 import coz.weavon.common.presentation.model.validation.ValidPagination;
 import jakarta.validation.ConstraintValidator;
