+++ src/main/java/coz/weavon/context/member/application/model/command/MemberSearchCommand.java
@@ -1,6 +1,6 @@
 package coz.weavon.context.member.application.model.command;
 
-import coz.weavon.common.application.model.RestCommand;
+import coz.weavon.common.application.model.command.RestCommand;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
